[AudioProcessor] add PreprocessingMixin
Regenerated from audio-processor by split_stack.py -- do not amend by hand;
edit the monolith and re-run the script instead. PR #48110.
